+++ src/components/pages/get-a-quote.svelte
@@ -3,8 +3,8 @@
   import type { GetAQuotePageStoryblok } from '$types/bloks';
   import Form from './get-a-quote/form.svelte';
   import Images from '../images.svelte';
-  import Steps from './get-a-quote/steps.svelte';
   import Timezone from './get-a-quote/timezone.svelte';
+  import Steps from '$components/blocks/steps.svelte';
 
   export let page: GetAQuotePageStoryblok;
 </script>
@@ -12,5 +12,13 @@
 <Seo />
 <Form {page} />
 <Timezone {page} />
-<Steps {page} variant="get-a-quote" />
+<Steps
+  block={{
+    _uid: 'GetAQuoteSteps',
+    component: 'steps',
+    steps: page.steps,
+    steps_subtitle: page.steps_subtitle,
+    steps_title: page.title
+  }}
+/>
 <Images images={page.images} class="py-12 md:py-20 lg:py-32" />
