+++ src/components/pages/home-page.svelte
@@ -14,12 +14,12 @@
   import { Button } from '@significa/svelte-ui';
   import Seo from '$components/seo.svelte';
   import { afterNavigate } from '$app/navigation';
-  import Testimonials from '$components/testimonials.svelte';
   import { getAnchorFromCmsLink, getImageAttributes } from '$lib/utils/cms';
   import Services from './home/services.svelte';
   import { TrackingEvent, track } from '$lib/track';
   import { drawerLinks } from '$lib/actions/drawer-links';
   import { t } from '$lib/i18n';
+  import Testimonials from '$components/blocks/testimonials.svelte';
 
   export let data: HomePageStoryblok;
   export let posts: BlogPostPage[] | undefined;
@@ -84,12 +84,17 @@
   <Services {data} />
 
   <Testimonials
-    variant="two"
-    firstTitle={data.testimonials_title1}
-    secondTitle={data.testimonials_title2}
-    testimonials={data.testimonials}
-    ctaLabel={data.testimonials_cta_label}
-    ctaLink={data.testimonials_cta_link}
+    block={{
+      _uid: 'HomeTestimonials',
+      component: 'testimonials',
+      testimonials: data.testimonials,
+      testimonials_cta_label: data.testimonials_cta_label,
+      testimonials_cta_link: data.testimonials_cta_link,
+      testimonials_title1: data.testimonials_title1,
+      testimonials_title2: data.testimonials_title2,
+      variant: data.variant,
+      size: data.size
+    }}
   />
 
   <section class="mt-16 border-y lg:mt-20">
